backend/requirements.txt
fastapi
uvicorn
pytest
factory_boy
sqlalchemy